- cmd: singularise params - scrape: handle different var/window initialization. Get ChannelId smarterly.

+++ App/YtReader/Search/YtSearch.cs
@@ -350,7 +350,7 @@ order by updated"; // always order by updated so that if sync fails, we can resu
   public class EsChannel : EsChannelTitle, IHasUpdated {
     [Keyword] public string    main_channel_id                       { get; set; }
     [Keyword] public string    logo_url                              { get; set; }
-    public           decimal?  relevance                             { get; set; }
+    public           double?  relevance                             { get; set; }
     public           long?     subs                                  { get; set; }
     public           long?     channel_views                         { get; set; }
     [Keyword] public string    country                               { get; set; }
